% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{bcf_getContigs}
\alias{bcf_getContigs}
\title{Reads contig names and lengths from BCF file as a data frame.}
\usage{
bcf_getContigs(path)
}
\arguments{
\item{path}{Path to VCF or BCF file}
}
\description{
Reads contig names and lengths from BCF file as a data frame.
}
