% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{bcf_getGTandAD}
\alias{bcf_getGTandAD}
\title{Reads gentoype (0, 1, 2, NA coded) and allele depths from BCF. Optionally,
reads only from a specific region of the reference.}
\usage{
bcf_getGTandAD(path, region = NULL, rowsAreSamples = T, minMAF = 0,
  maxMissing = 1)
}
\arguments{
\item{path}{Path to VCF or BCF file}

\item{region}{A region (like 'Chr1:1-1000') to extract. `path` must be an
indexed BCF if this option is used.}

\item{rowsAreSamples}{If TRUE, transpose GT and AD matricies so rows are
samples and SNPs are columns. This is the opposite of a BCF.}
}
\description{
Reads gentoype (0, 1, 2, NA coded) and allele depths from BCF. Optionally,
reads only from a specific region of the reference.
}
