% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{bcf_getSamples}
\alias{bcf_getSamples}
\title{Reads samples names from BCF/VCF}
\usage{
bcf_getSamples(path)
}
\arguments{
\item{path}{Path to VCF or BCF file}
}
\description{
Reads samples names from BCF/VCF
}
