% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBCF.R
\name{bcf_getWindows}
\alias{bcf_getWindows}
\title{Generate a list of regions which slide over the genome}
\usage{
bcf_getWindows(path, windowsize = 1000, slide = windowsize, chrom = NULL,
  from = NULL, to = NULL)
}
\arguments{
\item{path}{Path to VCF or BCF file}

\item{windowsize}{Size of each window, in bases}

\item{slide}{Step size of each window, in bases}

\item{chrom}{Only generate windows on contig 'chrom'}

\item{from}{Only generate windows starting at `from` on `chrom`}

\item{to}{Only generate windows until `to` on `chrom`}
}
\description{
Generate a list of regions which slide over the genome
}
