% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{window_halfmax}
\alias{window_halfmax}
\title{window_halfmax}
\usage{
window_halfmax(bcf, region, minMAF = 0, maxMissing = 1, geno = NULL,
  samples = NULL)
}
\arguments{
\item{bcf}{Indexed BCF or VCF file}

\item{region}{Region of genome}

\item{minMAF}{Minimum SNP minor allle freq}

\item{maxMissing}{Maximum SNP missing data rate}

\item{samples}{Set of samples to calculate LD among}
}
\value{
data.frame with one row and columns rho, halfmax, region, nsnp
}
\description{
window_halfmax
}
