% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{windowed_halfmax}
\alias{windowed_halfmax}
\title{windowed_halfmax}
\usage{
windowed_halfmax(bcf, windowsize, slide = windowsize, minMAF = 0.1,
  maxMissing = 0.8, windows = NULL)
}
\arguments{
\item{bcf}{Indexed BCF or VCF file}

\item{windowsize}{Size of each window}

\item{slide}{Number of bases to skip between each window start}

\item{minMAF}{Minimum SNP minor allle freq}

\item{maxMissing}{Maximum SNP missing data rate}

\item{windows}{restrict analyses to these windows (expected to be created with bcf_getWindows)}
}
\value{
data.frame  columns rho, halfmax, region, nsnp
}
\description{
windowed_halfmax
}
