% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{windowed_halfmax}
\alias{windowed_halfmax}
\title{windowed_halfmax}
\usage{
windowed_halfmax(bcf, windowsize = NULL, slide = windowsize, minMAF = 0.1,
  maxMissing = 0.8, windows = NULL, chrom = NULL, from = NULL,
  to = NULL, errors = "stop")
}
\arguments{
\item{bcf}{Indexed BCF or VCF file}

\item{windowsize}{Size of each window (required unless `windows` provided)}

\item{slide}{Number of bases to skip between each window start (default `windowsize`)}

\item{minMAF}{Minimum SNP minor allle freq}

\item{maxMissing}{Maximum SNP missing data rate}

\item{windows}{restrict analyses to these windows (expected to be created with bcf_getWindows)}

\item{chrom}{Only generate windows on contig 'chrom'}

\item{from}{Only generate windows starting at `from` on `chrom`}

\item{to}{Only generate windows until `to` on `chrom`}

\item{errors}{One of "remove" or "stop", see ?foreach::foreach 's .errorhandling parameter}
}
\value{
data.frame  columns rho, halfmax, region, nsnp
}
\description{
windowed_halfmax
}
